Added Rumout Buffer and Universal Box with MMB V2 compatability.
Formerly Cotton Tail Lite and pregate switches, the renamed runout buffer is what makes endless spool, runout and pre loading work on the ERCF.

Third party buffers should also be paired with this to ensure full compatibility with Happy Hare.

This revision has had a small increase in runout buffer to 14.6mm which will increase the tolerance and reliability of endless spool.

All standard V2 kits have the required components for this so no additional hardware will be necessary.