Development edition of Filament Path with Sensor switch
Sensor Switch Filament Path Block

This is the micro switch edition of the filament path block. This now adopt sthe use of levered switches, MR85 bearings, and should replaces the need for the runout buffer for Filamentalist.

This should only be used as a dual paralell switch configuration for ERCT or for Filamentalist. It is not reliable for endless spool runout. If a thirdparty buffer is used (pika, quiver, etc....) the runout buffer should be used instead of this option.

Wiring of this is recommended to be soldered at right angles and inline with the filament path, with a pigtail header wire so it's easy to connect.