**C10 Filament Block**
Change Log

- Gears have been relocated and optimized in the block design to be spaced optimally for correct filament pressure
- Idler gear is now integrated into the filament body reducing printing variations/errors that were stacked in the previous design.
- Reduced interface points. Elimination of catch points with the elimination of brake/trap holes.
- Reduced wide openings. Giving far better stability to the filament so it does not miss the encoder hole.
- Elimination of the old brake/trap. This has now been removed and relocated.
- Redesigned 2 part Latch design. Latch now acts as a 'parking' brake for preventing filament from drifting. 2nd part is replaceable, and secured with a M2x10 SHCS
- Redesigned top hat with gear positions. Still a traditional three (or can be used as a two gear system). Park, Neutral, and Drive. Top hat now snaps in place.
- New hinge design. Removed the floating hinge design from the earlier designs.
- relocated the the Base screw to the middle hole with a new easier to access magnet holder to service it.
- New clamp adjustment screw (requires a M3x6mm or 8mm FHCS - provides better control over too much filament grip and squish.

The only thing that changes for most people in terms of BOM is getting FHCS fasteners. These are not required, and you can make it work without it. It's meant to provide a stop so you don't squish things too far. Calibartion of this is similar to how it's done on an Stealthburner where you adjust it and find the spot that it barely holds onto the filament. This is important to do because if you have it set to far down, it can carve filament on both sides excessively, especially on any system that is set up for high speed.

The servo positions will need to be recalibrated to the new three index points on the top-hat.